Add new restaurant activity/ update list name
Added new restaurant activity.

Floating action button on edit list page opens
new restaurant activity.

Can now update the list name in the edit list
activity.

+++ app/src/main/java/com/erw/randomrestaurant/database/Restaurant.java
@@ -27,7 +27,7 @@ public class Restaurant implements Parcelable {
     public Restaurant(){}
 
     @Ignore
-    public Restaurant(int listId, String name){
+    public Restaurant(long listId, String name){
         this.listId = listId;
         this.name = name;
     }
